refactor(assistants): card grid for My Assistants, single ordering surface
Consolidate assistant ordering to the Enabled tab and drop the competing
drag-reorder on the My Assistants tab:

- My Assistants now renders a card grid (mirrors the Official cards) instead
  of a draggable row list; Created-by-you group on top, Your CLI below.
- Remove dnd-kit drag/reorder from My Assistants (MyAssistantRow deleted,
  onReorder wiring removed from AssistantHomeTabs and the settings page).
- Update the My Assistants hint copy across all locales (drop 'drag to reorder').

Ordering now lives in exactly one place (the Enabled tab).
