+++ packages/desktop/src/renderer/pages/settings/AssistantSettings/home/MyAssistantsList.tsx
@@ -4,16 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { DragEndEvent } from '@dnd-kit/core';
 import type { AssistantListItem } from '../types';
 import { type AssistantEnabledFilter, filterByEnabled, groupMyAssistants } from '../assistantUtils';
-import MyAssistantRow from './MyAssistantRow';
-import { DndContext, PointerSensor, closestCenter, useSensor, useSensors } from '@dnd-kit/core';
-import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable';
+import MyAssistantCard from './MyAssistantCard';
 import { useTalkToButler } from '@/renderer/hooks/assistant/useTalkToButler';
 import { Dropdown, Menu, Button } from '@arco-design/web-react';
-import { Down, SortTwo } from '@icon-park/react';
-import React, { useCallback, useMemo, useState } from 'react';
+import { AllApplication, Down } from '@icon-park/react';
+import React, { useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 type MyAssistantsListProps = {
@@ -22,7 +19,6 @@ type MyAssistantsListProps = {
   onOpenDetail: (assistant: AssistantListItem) => void;
   onDelete: (assistant: AssistantListItem) => void;
   onToggleEnabled: (assistant: AssistantListItem, checked: boolean) => void;
-  onReorder: (activeId: string, overId: string) => void | Promise<void>;
   onStartChat: (assistant: AssistantListItem) => void;
   /** Switch to the Official tab (to duplicate an official assistant). */
   onGoOfficial: () => void;
@@ -31,13 +27,22 @@ type MyAssistantsListProps = {
 
 const FILTER_OPTIONS: AssistantEnabledFilter[] = ['all', 'enabled', 'disabled'];
 
+const renderGroupHeader = (title: string, count: number, barClass: string) => (
+  <div className='mb-10px flex items-center gap-8px px-2px'>
+    <span className={`h-13px w-3px rounded-2px ${barClass}`} />
+    <span className='text-12px font-600 text-t-secondary'>{title}</span>
+    {count > 0 ? (
+      <span className='rounded-999px bg-fill-2 px-6px py-1px text-10px font-500 text-t-quaternary'>{count}</span>
+    ) : null}
+  </div>
+);
+
 const MyAssistantsList: React.FC<MyAssistantsListProps> = ({
   assistants,
   localeKey,
   onOpenDetail,
   onDelete,
   onToggleEnabled,
-  onReorder,
   onStartChat,
   onGoOfficial,
   searchActive = false,
@@ -55,25 +60,11 @@ const MyAssistantsList: React.FC<MyAssistantsListProps> = ({
       }),
     });
   };
-  const sensors = useSensors(useSensor(PointerSensor, { activationConstraint: { distance: 8 } }));
-
-  // Drag reorder is only meaningful in the unfiltered "all" view; a filtered
-  // view hides rows, so dragging would produce an ambiguous global order.
-  const draggable = filter === 'all' && !searchActive;
 
   const { cliAssistants, createdAssistants } = useMemo(() => {
     const filtered = filterByEnabled(assistants, filter);
     return groupMyAssistants(filtered);
   }, [assistants, filter]);
-
-  const handleDragEnd = useCallback(
-    (event: DragEndEvent) => {
-      const { active, over } = event;
-      if (!draggable || !over || active.id === over.id) return;
-      void onReorder(String(active.id), String(over.id));
-    },
-    [draggable, onReorder]
-  );
 
   const filterMenu = (
     <Menu onClickMenuItem={(key) => setFilter(key as AssistantEnabledFilter)}>
@@ -87,38 +78,21 @@ const MyAssistantsList: React.FC<MyAssistantsListProps> = ({
     </Menu>
   );
 
-  const renderGroup = (title: string, list: AssistantListItem[], testId: string, barClass: string) => {
-    if (list.length === 0) return null;
-    return (
-      <div className='mt-20px first:mt-0' data-testid={testId}>
-        <div className='mb-10px flex items-center gap-8px px-2px'>
-          <span className={`h-13px w-3px rounded-2px ${barClass}`} />
-          <span className='text-12px font-600 text-t-secondary'>{title}</span>
-          <span className='rounded-999px bg-fill-2 px-6px py-1px text-10px font-500 text-t-quaternary'>
-            {list.length}
-          </span>
-        </div>
-        <DndContext sensors={sensors} collisionDetection={closestCenter} onDragEnd={handleDragEnd}>
-          <SortableContext items={list.map((a) => a.id)} strategy={verticalListSortingStrategy}>
-            <div className='space-y-8px'>
-              {list.map((assistant) => (
-                <MyAssistantRow
-                  key={assistant.id}
-                  assistant={assistant}
-                  localeKey={localeKey}
-                  draggable={draggable}
-                  onOpenDetail={onOpenDetail}
-                  onDelete={onDelete}
-                  onToggleEnabled={onToggleEnabled}
-                  onStartChat={onStartChat}
-                />
-              ))}
-            </div>
-          </SortableContext>
-        </DndContext>
-      </div>
-    );
-  };
+  const renderCardGrid = (list: AssistantListItem[]) => (
+    <div className='grid grid-cols-1 gap-14px sm:grid-cols-2 lg:grid-cols-3'>
+      {list.map((assistant) => (
+        <MyAssistantCard
+          key={assistant.id}
+          assistant={assistant}
+          localeKey={localeKey}
+          onOpenDetail={onOpenDetail}
+          onDelete={onDelete}
+          onToggleEnabled={onToggleEnabled}
+          onStartChat={onStartChat}
+        />
+      ))}
+    </div>
+  );
 
   // The "created by me" group shows a guiding empty state when the user has
   // no custom assistants yet (only in the unfiltered view — a filtered empty
@@ -158,11 +132,10 @@ const MyAssistantsList: React.FC<MyAssistantsListProps> = ({
 
   return (
     <div data-testid='my-assistants-pane'>
-      {/* Compact toolbar: a quiet reorder hint (icon + tooltip) on the left, the
-          enabled filter on the right — no full-width banner hogging a row. */}
-      <div className='mb-4px flex items-center justify-between'>
+      {/* Compact toolbar: quiet one-line hint (full text on hover) + enabled filter. */}
+      <div className='mb-14px flex items-center justify-between gap-12px'>
         <span className='inline-flex min-w-0 items-center gap-6px text-12px text-t-tertiary'>
-          <SortTwo
+          <AllApplication
             theme='outline'
             size={14}
             fill='currentColor'
@@ -171,7 +144,7 @@ const MyAssistantsList: React.FC<MyAssistantsListProps> = ({
           />
           <span className='truncate'>
             {t('settings.myAssistantsHintShort', {
-              defaultValue: 'Your own assistants — used wherever you pick one. Drag to reorder.',
+              defaultValue: 'Your own assistants — used wherever you pick one.',
             })}
           </span>
         </span>
@@ -179,7 +152,7 @@ const MyAssistantsList: React.FC<MyAssistantsListProps> = ({
           <Button
             size='mini'
             data-testid='assistant-enabled-filter'
-            className='!flex !items-center !gap-4px !rounded-8px'
+            className='!flex !shrink-0 !items-center !gap-4px !rounded-8px'
           >
             <span>
               {t(`settings.assistantFilter.${filter}`, {
@@ -197,48 +170,23 @@ const MyAssistantsList: React.FC<MyAssistantsListProps> = ({
         </div>
       ) : null}
 
-      {renderGroup(
-        t('settings.assistantGroupCli', { defaultValue: 'Your CLI' }),
-        cliAssistants,
-        'group-cli',
-        'bg-warning-5'
-      )}
-
+      {/* Created-by-you group on top. */}
       {createdAssistants.length > 0 || createdEmpty ? (
-        <div className='mt-20px' data-testid='group-created-section'>
-          <div className='mb-10px flex items-center gap-8px px-2px'>
-            <span className='h-13px w-3px rounded-2px bg-primary-5' />
-            <span className='text-12px font-600 text-t-secondary'>
-              {t('settings.assistantGroupCreated', { defaultValue: 'Created by you' })}
-            </span>
-            {createdAssistants.length > 0 ? (
-              <span className='rounded-999px bg-fill-2 px-6px py-1px text-10px font-500 text-t-quaternary'>
-                {createdAssistants.length}
-              </span>
-            ) : null}
-          </div>
-          {createdEmpty ? (
-            renderCreatedEmpty()
-          ) : (
-            <DndContext sensors={sensors} collisionDetection={closestCenter} onDragEnd={handleDragEnd}>
-              <SortableContext items={createdAssistants.map((a) => a.id)} strategy={verticalListSortingStrategy}>
-                <div className='space-y-8px'>
-                  {createdAssistants.map((assistant) => (
-                    <MyAssistantRow
-                      key={assistant.id}
-                      assistant={assistant}
-                      localeKey={localeKey}
-                      draggable={draggable}
-                      onOpenDetail={onOpenDetail}
-                      onDelete={onDelete}
-                      onToggleEnabled={onToggleEnabled}
-                      onStartChat={onStartChat}
-                    />
-                  ))}
-                </div>
-              </SortableContext>
-            </DndContext>
+        <div data-testid='group-created-section'>
+          {renderGroupHeader(
+            t('settings.assistantGroupCreated', { defaultValue: 'Created by you' }),
+            createdAssistants.length,
+            'bg-primary-5'
           )}
+          {createdEmpty ? renderCreatedEmpty() : renderCardGrid(createdAssistants)}
+        </div>
+      ) : null}
+
+      {/* CLI group below. */}
+      {cliAssistants.length > 0 ? (
+        <div className='mt-20px' data-testid='group-cli'>
+          {renderGroupHeader(t('settings.assistantGroupCli', { defaultValue: 'Your CLI' }), cliAssistants.length, 'bg-warning-5')}
+          {renderCardGrid(cliAssistants)}
         </div>
       ) : null}
     </div>
