+++ packages/desktop/src/renderer/services/i18n/locales/de-DE/settings.json
@@ -162,7 +162,7 @@
   "assistantReorderHint": "Zum Umsortieren ziehen — das bestimmt die Anzeigereihenfolge überall dort, wo du einen Assistenten auswählst (Startseite, Teams, geplante Aufgaben…).",
   "assistantReorderHintShort": "Zum Umsortieren ziehen",
   "assistantReorderSearchDisabled": "Leere die Suche, um die Reihenfolge zu ändern.",
-  "myAssistantsHintShort": "Assistenten, die automatisch aus deinen lokalen CLIs erzeugt wurden, sowie selbst erstellte oder von offiziellen duplizierte. Ziehe sie, um die Anzeigereihenfolge zu ändern.",
+  "myAssistantsHintShort": "Assistenten, die automatisch aus deinen lokalen CLIs erzeugt wurden, sowie selbst erstellte oder von offiziellen duplizierte.",
   "assistantGroupCli": "CLI-Assistenten",
   "assistantGroupCreated": "Benutzerdefinierte Assistenten",
   "myAssistantsEmpty": "Hier gibt es noch keine Assistenten. Aktiviere einen offiziellen Assistenten oder verbinde ein lokales CLI-Tool.",
