+++ packages/desktop/src/renderer/services/i18n/locales/fa-IR/settings.json
@@ -159,7 +159,7 @@
   "assistantHomeLeadShort": "متخصصان هوش مصنوعی آماده کار، از قبل مجهز به مهارت‌ها و ابزارها. آن‌ها را فعال کنید تا در چت‌ها، تیم‌ها و کارهای زمان‌بندی‌شده استفاده شوند.",
   "assistantReorderHintShort": "برای تغییر ترتیب بکشید",
   "assistantReorderSearchDisabled": "برای تغییر ترتیب، جستجو را پاک کنید.",
-  "myAssistantsHintShort": "دستیارهایی که به‌صورت خودکار از CLIهای محلی شما ساخته شده‌اند، همراه با دستیارهایی که ساخته‌اید یا از نمونه‌های رسمی تکثیر کرده‌اید. برای تغییر ترتیب نمایش آن‌ها بکشید.",
+  "myAssistantsHintShort": "دستیارهایی که به‌صورت خودکار از CLIهای محلی شما ساخته شده‌اند، همراه با دستیارهایی که ساخته‌اید یا از نمونه‌های رسمی تکثیر کرده‌اید.",
   "assistantRuntimeLabel": "ایجنت:",
   "assistantGoChat": "چت",
   "assistantReorderHint": "برای تغییر ترتیب بکشید؛ این ترتیب مشخص می‌کند دستیارها در محل‌های انتخاب دستیار چگونه نمایش داده شوند (خانه، تیم‌ها، کارهای زمان‌بندی‌شده و ...).",
