+++ packages/desktop/src/renderer/services/i18n/locales/ko-KR/settings.json
@@ -162,7 +162,7 @@
   "assistantReorderHint": "드래그하여 순서를 바꿀 수 있습니다. 이 순서는 어시스턴트를 선택하는 모든 곳(홈, 팀, 예약 작업 등)에서의 표시 순서를 결정합니다.",
   "assistantReorderHintShort": "드래그하여 순서 변경",
   "assistantReorderSearchDisabled": "검색을 지우면 순서를 변경할 수 있습니다.",
-  "myAssistantsHintShort": "로컬 CLI에서 자동 생성되었거나 직접 만들었거나 공식에서 복제한 어시스턴트입니다. 드래그하여 표시 순서를 조정할 수 있습니다.",
+  "myAssistantsHintShort": "로컬 CLI에서 자동 생성되었거나 직접 만들었거나 공식에서 복제한 어시스턴트입니다.",
   "assistantGroupCli": "CLI 어시스턴트",
   "assistantGroupCreated": "사용자 지정 어시스턴트",
   "myAssistantsEmpty": "아직 어시스턴트가 없습니다. 공식 어시스턴트를 활성화하거나 로컬 CLI 도구를 연결하세요.",
