+++ packages/desktop/src/renderer/services/i18n/locales/pt-BR/settings.json
@@ -162,7 +162,7 @@
   "assistantReorderHint": "Arraste para reordenar — isso define a ordem de exibição em todos os lugares onde você escolhe um assistente (início, equipes, tarefas agendadas…).",
   "assistantReorderHintShort": "Arraste para reordenar",
   "assistantReorderSearchDisabled": "Limpe a busca para reordenar.",
-  "myAssistantsHintShort": "Assistentes gerados automaticamente das suas CLIs locais, além dos que você criou ou duplicou dos oficiais. Arraste para reordenar como eles aparecem.",
+  "myAssistantsHintShort": "Assistentes gerados automaticamente das suas CLIs locais, além dos que você criou ou duplicou dos oficiais.",
   "assistantGroupCli": "Assistentes de CLI",
   "assistantGroupCreated": "Assistentes personalizados",
   "myAssistantsEmpty": "Ainda não há assistentes aqui. Ative um assistente oficial ou conecte uma ferramenta CLI local.",
