+++ packages/desktop/src/renderer/services/i18n/locales/ru-RU/settings.json
@@ -162,7 +162,7 @@
   "assistantReorderHint": "Перетащите, чтобы изменить порядок — он определяет порядок отображения везде, где вы выбираете ассистента (главная, команды, запланированные задачи…).",
   "assistantReorderHintShort": "Перетащите для сортировки",
   "assistantReorderSearchDisabled": "Очистите поиск, чтобы изменить порядок.",
-  "myAssistantsHintShort": "Ассистенты, автоматически созданные из ваших локальных CLI, а также созданные вами или продублированные из официальных. Перетаскивайте, чтобы изменить порядок отображения.",
+  "myAssistantsHintShort": "Ассистенты, автоматически созданные из ваших локальных CLI, а также созданные вами или продублированные из официальных.",
   "assistantGroupCli": "CLI-ассистенты",
   "assistantGroupCreated": "Пользовательские ассистенты",
   "myAssistantsEmpty": "Здесь пока нет ассистентов. Включите официального ассистента или подключите локальный CLI-инструмент.",
