+++ packages/desktop/src/renderer/services/i18n/locales/tr-TR/settings.json
@@ -162,7 +162,7 @@
   "assistantReorderHint": "Sıralamak için sürükleyin — bu, bir asistan seçtiğiniz her yerde (ana ekran, ekipler, zamanlanmış görevler…) görüntülenme sırasını belirler.",
   "assistantReorderHintShort": "Sıralamak için sürükleyin",
   "assistantReorderSearchDisabled": "Sıralamak için aramayı temizleyin.",
-  "myAssistantsHintShort": "Yerel CLI’lerinizden otomatik oluşturulan, kendi oluşturduğunuz veya resmi olanlardan çoğalttığınız asistanlar. Görünüm sırasını değiştirmek için sürükleyin.",
+  "myAssistantsHintShort": "Yerel CLI’lerinizden otomatik oluşturulan, kendi oluşturduğunuz veya resmi olanlardan çoğalttığınız asistanlar.",
   "assistantGroupCli": "CLI Asistanları",
   "assistantGroupCreated": "Özel Asistanlar",
   "myAssistantsEmpty": "Burada henüz asistan yok. Resmi bir asistanı etkinleştirin veya yerel bir CLI aracı bağlayın.",
