+++ packages/desktop/src/renderer/services/i18n/locales/uk-UA/settings.json
@@ -162,7 +162,7 @@
   "assistantReorderHint": "Перетягніть, щоб змінити порядок — він визначає порядок відображення всюди, де ви обираєте асистента (головна, команди, заплановані завдання…).",
   "assistantReorderHintShort": "Перетягніть для сортування",
   "assistantReorderSearchDisabled": "Очистьте пошук, щоб змінити порядок.",
-  "myAssistantsHintShort": "Асистенти, автоматично створені з ваших локальних CLI, а також створені вами чи продубльовані з офіційних. Перетягуйте, щоб змінити порядок відображення.",
+  "myAssistantsHintShort": "Асистенти, автоматично створені з ваших локальних CLI, а також створені вами чи продубльовані з офіційних.",
   "assistantGroupCli": "CLI-асистенти",
   "assistantGroupCreated": "Користувацькі асистенти",
   "myAssistantsEmpty": "Тут поки немає асистентів. Увімкніть офіційного асистента або підключіть локальний CLI-інструмент.",
