+++ packages/desktop/src/renderer/services/i18n/locales/zh-CN/settings.json
@@ -159,7 +159,7 @@
   "assistantHomeLeadShort": "配好技能和工具、能直接上岗的 AI 专家。启用后即可用于对话、开团和定时任务。",
   "assistantReorderHintShort": "拖拽排序",
   "assistantReorderSearchDisabled": "清除搜索后即可调整顺序。",
-  "myAssistantsHintShort": "由本机 CLI 自动生成、你自己创建或从官方复制的助手。拖拽可调整它们的展示顺序。",
+  "myAssistantsHintShort": "由本机 CLI 自动生成、你自己创建或从官方复制的助手。",
   "assistantRuntimeLabel": "Agent:",
   "assistantGoChat": "去对话",
   "assistantReorderHint": "拖拽可调整顺序，它决定你挑选助手时（首页、开团、定时任务等）列表里的展示先后。",
