+++ packages/desktop/src/renderer/services/i18n/locales/zh-TW/settings.json
@@ -162,7 +162,7 @@
   "assistantReorderHint": "拖曳可調整順序，它決定你挑選助理時（首頁、開團、定時任務等）清單裡的顯示先後。",
   "assistantReorderHintShort": "拖曳排序",
   "assistantReorderSearchDisabled": "清除搜尋後即可調整順序。",
-  "myAssistantsHintShort": "由本機 CLI 自動生成、你自己建立或從官方複製的助手。拖曳可調整它們的顯示順序。",
+  "myAssistantsHintShort": "由本機 CLI 自動生成、你自己建立或從官方複製的助手。",
   "assistantGroupCli": "CLI 助手",
   "assistantGroupCreated": "自訂助手",
   "myAssistantsEmpty": "你的助理清單還是空的。去「官方助理」啟用一個專家，或讓 AionUi 辨識你本機的 CLI 工具。",
